Add 'coverage' library for code measurement
This commit integrates the 'coverage' library version 7.3.2 into the project. This library is used for code coverage measurement in Python. It is a necessary addition to allow for tracking which parts of the code have been executed thus helping in identifying which branches of conditional statements have not been tested. In the GitHub workflows, commands for running coverage report and coverage lcov have been added. The format for the Coveralls GitHub Action was updated to 'lcov'. Additionally, the secret token for GitHub was added to the workflow.