Updated project dependencies and improved logging messages syntax for better Python version support
This commit modifies typed syntax and adjusts the project's dependencies for compatibility across Python versions 3.8 and up. The typed syntax relating to the asyncio VitalWorkDispatcher class, aiohttp web server, and API has been modified. The API also replaces f-strings for logging purposes with older string formatting %s for greater compatibility. The dependencies listed in pyproject.toml and poetry.lock have been updated, and the version of Python specified for the project changes from 3.8 < version < 3.12 to version >= 3.8. The project now supports the latest Python version explicitly.

Additionally, the package 'coverage' has been moved to the 'test' group, and 'tox' has been added as a dependency. The commit also introduces changes to the status, search, and fetch logging messages in api.py. The aim of these changes is to improve the interoperability of the project with various Python versions and to